Restructure quasi-Newton parallel tests
